requirements.txt
# Core RAG dependencies
langchain==0.3.23
langchain-community==0.3.21
langchain-openai==0.3.11

# Vector store
chromadb==0.6.3

# Embeddings
sentence-transformers==3.4.1

# Document loaders
pypdf==5.4.0
unstructured==0.17.2

# LLM provider
openai==1.68.2

# Environment variable management
python-dotenv==1.1.0
